Extract Windowed Serde into the public package API
