+++ streams/src/main/java/org/apache/kafka/streams/kstream/WindowedDeserializer.java
@@ -14,12 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.streams.kstream.internals;
+package org.apache.kafka.streams.kstream;
 
 import org.apache.kafka.common.config.ConfigException;
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.utils.Utils;
-import org.apache.kafka.streams.kstream.Windowed;
+import org.apache.kafka.streams.kstream.internals.UnlimitedWindow;
 
 import java.nio.ByteBuffer;
 import java.util.Map;
